# [Clojure] improvements