@jinyus 2023-10-17
This PR:

```
Clojure:

        Benchmark 1: java  -jar ./target/related.jar
        Processing time (w/o IO): 2708ms
        Processing time (w/o IO): 2808ms
        Processing time (w/o IO): 2802ms
        Processing time (w/o IO): 2792ms
        Processing time (w/o IO): 2691ms
        Processing time (w/o IO): 2754ms
        Processing time (w/o IO): 2699ms
        Processing time (w/o IO): 2800ms
        Processing time (w/o IO): 2766ms
        Processing time (w/o IO): 2769ms
        Processing time (w/o IO): 2783ms
        Processing time (w/o IO): 2717ms
        Processing time (w/o IO): 2749ms
          Time (mean ± σ):      3.763 s ±  0.048 s    [User: 5.826 s, System: 0.185 s]
          Range (min … max):    3.690 s …  3.835 s    10 runs
         
Clojure:

        Benchmark 1: java  -jar ./target/related.jar
        Processing time (w/o IO): 38528ms
        Processing time (w/o IO): 38545ms
        Processing time (w/o IO): 38987ms
          Time (mean ± σ):     39.984 s ±  0.327 s    [User: 42.360 s, System: 0.268 s]
          Range (min … max):   39.752 s … 40.215 s    2 runs
         
Clojure:

        Benchmark 1: java  -jar ./target/related.jar
        Processing time (w/o IO): 340448ms
        Processing time (w/o IO): 345973ms
        Processing time (w/o IO): 345256ms
          Time (mean ± σ):     347.417 s ±  0.512 s    [User: 350.683 s, System: 0.696 s]
          Range (min … max):   347.055 s … 347.779 s    2 runs

```

Main Branch:

```
Clojure:

        Benchmark 1: java  -jar ./target/related.jar
        Processing time (w/o IO): 4088ms
        Processing time (w/o IO): 3589ms
        Processing time (w/o IO): 4108ms
        Processing time (w/o IO): 3610ms
        Processing time (w/o IO): 3631ms
          Time (mean ± σ):      4.651 s ±  0.215 s    [User: 6.268 s, System: 0.162 s]
          Range (min … max):    4.473 s …  5.068 s    10 runs
         
Clojure:

        Benchmark 1: java  -jar ./target/related.jar
        Processing time (w/o IO): 50047ms
        Processing time (w/o IO): 51475ms
        Processing time (w/o IO): 58079ms
          Time (mean ± σ):     55.984 s ±  4.615 s    [User: 58.385 s, System: 0.264 s]
          Range (min … max):   52.721 s … 59.248 s    2 runs
         
Clojure:

        Benchmark 1: java  -jar ./target/related.jar
        Processing time (w/o IO): 506421ms
        Processing time (w/o IO): 439515ms
        Processing time (w/o IO): 439554ms
          Time (mean ± σ):     441.289 s ±  0.071 s    [User: 444.729 s, System: 0.592 s]
          Range (min … max):   441.239 s … 441.340 s    2 runs
```